# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: discovery
    description: Public companion discovery
  - name: dashboards
    description: Authenticated user workspace summaries
  - name: safety
    description: Safety resources and SafeSpots
  - name: bookings
    description: Booking lifecycle and server-calculated pricing
  - name: requests
    description: Structured Favor Requests (pre-payment)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /companions:
    get:
      operationId: listCompanions
      tags: [discovery]
      summary: List approved companions
      description: Returns approved companions using privacy-safe approximate areas only.
      parameters:
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: activity
          schema:
            type: string
        - in: query
          name: language
          schema:
            type: string
        - in: query
          name: maxRate
          schema:
            type: number
        - in: query
          name: instantBook
          schema:
            type: boolean
      responses:
        "200":
          description: Approved companion results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Companion"
  /companions/{id}:
    get:
      operationId: getCompanion
      tags: [discovery]
      summary: Get a companion profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Companion profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Companion"
        "404":
          description: Companion not found
  /dashboard/customer:
    get:
      operationId: getCustomerDashboard
      tags: [dashboards]
      summary: Get the authenticated customer's dashboard summary
      responses:
        "200":
          description: Customer dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDashboard"
  /dashboard/companion:
    get:
      operationId: getCompanionDashboard
      tags: [dashboards]
      summary: Get the authenticated companion's dashboard summary
      responses:
        "200":
          description: Companion dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanionDashboard"
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [dashboards]
      summary: Get live admin operations overview
      responses:
        "200":
          description: Admin overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
  /safety:
    get:
      operationId: getSafetyResources
      tags: [safety]
      summary: Get safety center resources
      responses:
        "200":
          description: Safety resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyResources"
  /safespots:
    get:
      operationId: listSafeSpots
      tags: [safety]
      summary: List public SafeSpots for a city
      parameters:
        - in: query
          name: city
          schema:
            type: string
      responses:
        "200":
          description: SafeSpot results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SafeSpot"
  /bookings/quote:
    get:
      operationId: getBookingQuote
      tags: [bookings]
      summary: Get a server-calculated price quote — never trust browser amounts
      parameters:
        - in: query
          name: companionId
          required: true
          schema:
            type: string
        - in: query
          name: durationHours
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Server-calculated price breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceQuote"
        "404":
          description: Companion not found
  /bookings:
    post:
      operationId: createBookingIntent
      tags: [bookings]
      summary: Create a booking intent with server-calculated pricing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking intent with full pricing breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          description: Authentication required
  /bookings/{id}/deposit:
    post:
      operationId: authorizeDeposit
      tags: [bookings]
      summary: Pay the refundable $10 deposit to unlock masked chat
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stripe Payment Intent client secret for the deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositResult"
        "401":
          description: Authentication required
        "404":
          description: Booking not found
  /bookings/{id}/authorize:
    post:
      operationId: authorizeFullPayment
      tags: [bookings]
      summary: Authorize full server-priced payment to confirm the booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stripe Payment Intent client secret for full amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositResult"
        "401":
          description: Authentication required
        "404":
          description: Booking not found
  /favor-requests:
    post:
      operationId: createFavorRequest
      tags: [requests]
      summary: Send a structured Favor Request to a companion — free, no chat yet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavorRequestInput"
      responses:
        "201":
          description: Favor Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavorRequest"
        "401":
          description: Authentication required
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Companion:
      type: object
      required:
        - id
        - displayName
        - city
        - serviceArea
        - activities
        - languages
        - hourlyRate
        - responseTime
        - rating
        - reviewCount
        - verified
        - instantBook
      properties:
        id:
          type: string
        displayName:
          type: string
        city:
          type: string
        serviceArea:
          type: string
          description: Approximate area, never an exact address
        activities:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
        responseTime:
          type: string
        rating:
          type: number
        reviewCount:
          type: number
          description: Whole-number review count
        verified:
          type: boolean
        instantBook:
          type: boolean
        biography:
          type: ["string", "null"]
        boundaries:
          type: array
          items:
            type: string
        photoUrl:
          type: ["string", "null"]
    PriceQuote:
      type: object
      description: >
        Server-calculated price breakdown. Never trust amounts from the browser.
        Example for a $100 favor: customer pays $105, companion receives $85, platform earns $20.
      required:
        - companionId
        - durationHours
        - subtotalCents
        - customerFeeCents
        - totalCents
        - companionPayoutCents
        - platformRevenueCents
        - customerFeePercent
        - companionCommissionPercent
        - depositCents
        - depositCreditedToFinal
      properties:
        companionId:
          type: string
        durationHours:
          type: number
        subtotalCents:
          type: number
          description: Companion hourly rate × duration in cents
        customerFeeCents:
          type: number
          description: 5% safety-and-service fee paid by the customer
        totalCents:
          type: number
          description: What the customer pays (subtotal + 5% fee)
        companionPayoutCents:
          type: number
          description: What the companion receives (subtotal minus 15% commission)
        platformRevenueCents:
          type: number
          description: Gross platform revenue (5% + 15% = 20% of subtotal)
        customerFeePercent:
          type: number
          description: Always 5
        companionCommissionPercent:
          type: number
          description: Always 15
        depositCents:
          type: number
          description: Refundable deposit to unlock chat — always 1000 (= $10)
        depositCreditedToFinal:
          type: boolean
          description: The $10 deposit is credited toward the final booking total
    CustomerDashboard:
      type: object
      required: [upcomingBookings, completedBookings, savedCompanions, safetyPlans]
      properties:
        upcomingBookings:
          type: number
        completedBookings:
          type: number
        savedCompanions:
          type: number
        safetyPlans:
          type: number
    CompanionDashboard:
      type: object
      required: [pendingRequests, upcomingBookings, earningsCents, profileViews]
      properties:
        pendingRequests:
          type: number
        upcomingBookings:
          type: number
        earningsCents:
          type: number
        profileViews:
          type: number
    AdminOverview:
      type: object
      required: [verificationQueue, openReports, activeBookings, checkInsDue]
      properties:
        verificationQueue:
          type: number
        openReports:
          type: number
        activeBookings:
          type: number
        checkInsDue:
          type: number
    SafetyResources:
      type: object
      required: [title, emergencyGuidance, principles]
      properties:
        title:
          type: string
        emergencyGuidance:
          type: string
        principles:
          type: array
          items:
            type: string
    SafeSpot:
      type: object
      required: [id, name, category, city, addressHint, openLate]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        city:
          type: string
        addressHint:
          type: string
        openLate:
          type: boolean
    BookingInput:
      type: object
      required: [companionId, activity, date, startTime, durationHours, safeSpotId]
      properties:
        companionId:
          type: string
        activity:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        durationHours:
          type: number
        safeSpotId:
          type: string
    Booking:
      type: object
      required: [id, status, subtotalCents, customerFeeCents, totalCents, companionPayoutCents, platformRevenueCents, depositCents, depositCreditedToFinal]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [draft, requested, deposit_paid, authorized, confirmed, completed, cancelled]
        subtotalCents:
          type: number
        customerFeeCents:
          type: number
        totalCents:
          type: number
        companionPayoutCents:
          type: number
        platformRevenueCents:
          type: number
        depositCents:
          type: number
        depositCreditedToFinal:
          type: boolean
        stripePaymentIntentId:
          type: ["string", "null"]
    DepositResult:
      type: object
      required: [bookingId, amountCents, clientSecret, creditedToFinal]
      properties:
        bookingId:
          type: string
        amountCents:
          type: number
        clientSecret:
          type: string
          description: Stripe Payment Intent client secret — only returned to the authenticated customer
        creditedToFinal:
          type: boolean
    FavorRequestInput:
      type: object
      required: [companionId, activity, preferredDate, preferredDurationHours]
      properties:
        companionId:
          type: string
        activity:
          type: string
        preferredDate:
          type: string
          format: date
        preferredDurationHours:
          type: number
        locationType:
          type: ["string", "null"]
          description: General location type e.g. café, museum, park — not a home address
        accessibilityNeeds:
          type: ["string", "null"]
        dressCode:
          type: ["string", "null"]
        additionalQuestions:
          type: ["string", "null"]
          description: Structured questions only — no contact info, personal data, or off-platform contact
    FavorRequest:
      type: object
      required: [id, status, companionId, activity, preferredDate, preferredDurationHours]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, accepted, declined, expired]
        companionId:
          type: string
        activity:
          type: string
        preferredDate:
          type: string
        preferredDurationHours:
          type: number
        createdAt:
          type: string