# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: string;
  status: BookingStatus;
  subtotalCents: number;
  customerFeeCents: number;
  totalCents: number;
  companionPayoutCents: number;
  platformRevenueCents: number;
  depositCents: number;
  depositCreditedToFinal: boolean;
  /** @nullable */
  stripePaymentIntentId?: string | null;
}