# lib/api-zod/src/generated/types/depositResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DepositResult {
  bookingId: string;
  amountCents: number;
  /** Stripe Payment Intent client secret — only returned to the authenticated customer */
  clientSecret: string;
  creditedToFinal: boolean;
}